build: make gosec happy by handling un-handled errors
