+++ examples/serialization/main.go
@@ -21,11 +21,11 @@ func main() {
 
 	// R1CS implements io.WriterTo and io.ReaderFrom
 	var buf bytes.Buffer
-	_r1cs.WriteTo(&buf)
+	_, _ = _r1cs.WriteTo(&buf)
 
 	// gnark objects (R1CS, ProvingKey, VerifyingKey, Proof) must be instantiated like so:
 	newR1CS := groth16.NewCS(ecc.BN254)
-	newR1CS.ReadFrom(&buf)
+	_, _ = newR1CS.ReadFrom(&buf)
 
 	// setup
 	pk, vk, _ := groth16.Setup(_r1cs)
@@ -34,15 +34,15 @@ func main() {
 	// groth16.ProvingKey, groth16.VerifyingKey and groth16.Proof implements io.WriterTo and io.ReaderFrom
 	// but also gnarkio.WriterRawTo to serialize without point compression
 	buf.Reset()
-	pk.WriteRawTo(&buf)
+	_, _ = pk.WriteRawTo(&buf)
 	newPK := groth16.NewProvingKey(ecc.BN254)
-	newPK.ReadFrom(&buf)
+	_, _ = newPK.ReadFrom(&buf)
 
 	// library user is free to use another encoder like gob or cbor to serialize / deserialize objects
 	// but will need to check reconstructed object validity (points on curve, etc)
 	// TODO add example and APIs for that
 	buf.Reset()
 	enc := cbor.NewEncoder(&buf)
-	enc.Encode(vk)
+	_ = enc.Encode(vk)
 	// ...
 }
