+++ internal/backend/bls24-315/plonk/verify.go
@@ -179,7 +179,9 @@ func Verify(proof *Proof, vk *VerifyingKey, publicWitness bls24_315witness.Witne
 		l, r, rl, o, one, // first part
 		_s1, _s2, // second & third part
 	}
-	linearizedPolynomialDigest.MultiExp(points, scalars, ecc.MultiExpConfig{ScalarsMont: true})
+	if _, err := linearizedPolynomialDigest.MultiExp(points, scalars, ecc.MultiExpConfig{ScalarsMont: true}); err != nil {
+		return err
+	}
 
 	// Fold the first proof
 	foldedProof, foldedDigest, err := kzg.FoldProof([]kzg.Digest{
